Add UI; Create containers for different sections
Added a UI consisting of a button for each option
as well as a result output and counters in order
to improve user interaction in the game. Added DIV
containers in the index file in order to seperate
different sections of the game as well as to make
it easier to add css to the game in the future.